Scenarios and several tests
Wrote down several test scenarios in a TestScenarios.md file.

Made a test that checks the filling of the calculator fields.

Made a test that fills fields, clicks to calculate and checks the calculated value.

Added method that finds the maximum input value and method that counts the number of characters.
TODO: move it to separate class and continue with implementing the test scenarios.