Input errors validation test and scenarios added
Implemented a class that takes XPath for the input field and contains methods:
to set values in field;
to check for validation errors present when entering values;
to get an error message.

Performed a test for the input fields, which checks for errors in cases 1, 0, -1.

Added scenarios into TestScenarios.md